knowledge/coreapi/core/requirements.txt
MySQL-python==1.2.5
Django==1.7
django-test-pep8==0.1
django-maintenancemode==0.10
pillow
djangorestframework==2.3.12
markdown
django-filter
PyJWT==0.1.6
selenium==2.39.0
django-rest-swagger==0.3.10
django-cors-headers
mock
gdata
django-sslserver
boto
django-storages==1.1.8
django-ses
autopep8
celery==3.1.25